// src/typings/types.ts
export type minimalType = {
    backgroundImage?: string;
    avatar: string;
    name: string;
    nameColor?: string;
    message: string;
    messageColor?: string;
    type?: string;
    typeColor?: string;
    circleBorder?: boolean;
}

export type panoramaType = {
    backgroundImage?: string;
    avatar: string;
    name: string;
    nameColor?: string;
    type?: string;
    typeColor?: string;
    circleBorder?: boolean;
}